Testing findings and integration
- Added an api to return size of checksum.
- Moved the client code to lib folder.
- add option in data corruption tool to pass parsing limit and starting
  offset while reading metadata file
- Moved code from misc.c to cksum_utils.c
- RPC split changes
