+++ ioservice/io_foms.c
@@ -27,6 +27,7 @@
 #include "lib/assert.h"
 #include "lib/misc.h"    /* M0_BITS */
 #include "lib/finject.h"
+#include "lib/cksum_utils.h"
 #include "net/net_internal.h"
 #include "net/buffer_pool.h"
 #include "fop/fop.h"
@@ -2162,7 +2163,7 @@ static int stob_io_create(struct m0_fom *fom)
 	 * checksum-nob for all stobs 
 	 */
 	M0_ASSERT( m0_is_read_fop(fom->fo_fop) ? 
-				(curr_cksum_nob == rw_replyfop->rwr_di_data_cksum.b_nob) : true);
+				(curr_cksum_nob == rw_replyfop->rwr_di_data_cksum.b_nob) : (curr_cksum_nob == rwfop->crw_di_data_cksum.b_nob));
 
 	if (rc != 0 && i > 0) {
 		while (--i >= 0) {
@@ -2429,6 +2430,7 @@ static void m0_io_fom_cob_rw_fini(struct m0_fom *fom)
 	if (fom_obj->fcrw_stio != NULL)
 		stob_io_destroy(fom);
 
+#if 0
 	if(m0_is_read_fop(fop))
 	{
 		struct m0_fop_cob_rw_reply *rw_replyfop = io_rw_rep_get(fom->fo_rep_fop);
@@ -2436,7 +2438,7 @@ static void m0_io_fom_cob_rw_fini(struct m0_fom *fom)
 		if(rw_replyfop->rwr_di_data_cksum.b_addr)
 			m0_buf_free(&rw_replyfop->rwr_di_data_cksum);
 	}
-	
+#endif	
 	m0_fom_fini(fom);
 	m0_free(fom_obj);
 }
