+++ motr/pg.h
@@ -809,6 +809,10 @@ struct target_ioreq {
 	// TODO: Combine this into one struct for checksums
 	struct m0_buf                  ti_attrbuf;
 	m0_bcount_t					   ti_cksum_copied;
+
+	/* Array for segment having b_nob value of checksum */
+	uint32_t                      *ti_cksum_seg_b_nob; 
+	
 	/**
 	 * Index vector containing segment number of attribute bufvec 
 	 * m0_op_io::ioo_attr for this target.
